refactor: move differ package to gitops package
