+++ cmd/swarm-deploy/main.go
@@ -17,7 +17,6 @@ import (
 	"github.com/swarm-deploy/swarm-deploy/internal/assistant"
 	"github.com/swarm-deploy/swarm-deploy/internal/config"
 	"github.com/swarm-deploy/swarm-deploy/internal/deployer"
-	"github.com/swarm-deploy/swarm-deploy/internal/differ"
 	"github.com/swarm-deploy/swarm-deploy/internal/entrypoints/healthserver"
 	"github.com/swarm-deploy/swarm-deploy/internal/entrypoints/mcpserver"
 	"github.com/swarm-deploy/swarm-deploy/internal/entrypoints/webhookserver"
@@ -31,6 +30,7 @@ import (
 	notify2 "github.com/swarm-deploy/swarm-deploy/internal/event/notify"
 	"github.com/swarm-deploy/swarm-deploy/internal/gitops/controller"
 	"github.com/swarm-deploy/swarm-deploy/internal/gitops/controller/statem"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ"
 	gitx "github.com/swarm-deploy/swarm-deploy/internal/gitops/git"
 	"github.com/swarm-deploy/swarm-deploy/internal/metrics"
 	swarmnode "github.com/swarm-deploy/swarm-deploy/internal/node"
