+++ internal/entrypoints/mcpserver/tools/git_commit_diff.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/artarts36/envmasker"
 	"github.com/swarm-deploy/swarm-deploy/internal/config"
-	"github.com/swarm-deploy/swarm-deploy/internal/differ"
-	"github.com/swarm-deploy/swarm-deploy/internal/differ/diff"
 	"github.com/swarm-deploy/swarm-deploy/internal/entrypoints/mcpserver/routing"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ/diff"
 	"github.com/swarm-deploy/swarm-deploy/internal/gitops/git"
 )
