+++ internal/entrypoints/mcpserver/tools/types.go
@@ -4,9 +4,9 @@ import (
 	"context"
 	"net"
 
-	"github.com/swarm-deploy/swarm-deploy/internal/differ"
-	"github.com/swarm-deploy/swarm-deploy/internal/differ/diff"
 	"github.com/swarm-deploy/swarm-deploy/internal/event/history"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ/diff"
 	gitx "github.com/swarm-deploy/swarm-deploy/internal/gitops/git"
 	"github.com/swarm-deploy/swarm-deploy/internal/registry"
 	"github.com/swarm-deploy/swarm-deploy/internal/service"
