+++ internal/gitops/differ/srvcomparator/helpers.go
