+++ internal/gitops/differ/srvcomparator/secret.go
@@ -4,7 +4,7 @@ import (
 	"sort"
 
 	"github.com/swarm-deploy/swarm-deploy/internal/compose"
-	"github.com/swarm-deploy/swarm-deploy/internal/differ/diff"
+	"github.com/swarm-deploy/swarm-deploy/internal/gitops/differ/diff"
 )
 
 type SecretComparator struct{}
